Feature: View my Surveys
- View my surveys shows just surveys created with an ID that matches the
  current users ID
- Moved Delete Survey into the view my Surveys
- Added description to create survey
- Moved view surveys and view my surveys into nav bar

+++ assets/scripts/survey/events.js
@@ -29,9 +29,11 @@ const onShowSurvey = function (event) {
 const onDeleteSurvey = function (event) {
   event.preventDefault()
   const data = getFormFields(event.target)
-  api.deleteSurvey(data.survey.id)
-    .then(ui.deleteSurveySuccess)
-    .catch(ui.deleteSurveyFailure)
+  if (confirm('Are you sure you want to delete this survey')) {
+    api.deleteSurvey(data.survey.id)
+      .then(ui.deleteSurveySuccess)
+      .catch(ui.deleteSurveyFailure)
+  }
 }
 const onUpdateSurvey = function (event) {
   event.preventDefault()
@@ -46,13 +48,22 @@ const hideContent = function () {
   $('.content').addClass('hidden')
 }
 
+const onUserSurveys = function () {
+  event.preventDefault()
+  api.indexSurvey()
+    .then(ui.userSurveysSuccess)
+    .catch(ui)
+}
+
 const addHandlers = function () {
   $('#create-survey').on('submit', onCreateSurvey)
   $('#index-survey').on('submit', onIndexSurvey)
   $('#show-survey').on('submit', onShowSurvey)
-  $('#delete-survey').on('submit', onDeleteSurvey)
+  // $('#delete-survey').on('submit', onDeleteSurvey)
+  $('.content').on('submit', '.delete-survey', onDeleteSurvey)
   $('#update-survey').on('submit', onUpdateSurvey)
   $('#clear-list').on('click', hideContent)
+  $('#user-surveys').on('submit', onUserSurveys)
 }
 
 module.exports = {
